Fixed some tests failing after migrating to Python 3

+++ qa/rpc-tests/test_framework/wsproxy.py
@@ -15,7 +15,7 @@ class JSONWSException(Exception):
 
 def EncodeDecimal(o):
     if isinstance(o, decimal.Decimal):
-        return round(o, 8)
+        return float(round(o, 8))
     raise TypeError(repr(o) + " is not JSON serializable")
 
 def __func():
